# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SORRISO PERFEITO API",
    "description": "API interativa para gestão de clínica odontológica, incluindo controle de acessos (usuários), pacientes, prontuário (anamnese), odontólogos, agendamento de consultas e catalogação de procedimentos clínicos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/odontologos": {
      "get": {
        "tags": [
          "Odontólogos"
        ],
        "summary": "Listar odontólogos cadastrados",
        "responses": {
          "200": {
            "description": "Lista de odontólogos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Odontologo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Odontólogos"
        ],
        "summary": "Cadastrar novo odontólogo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OdontologoInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Odontólogo cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "CRO já cadastrado"
          }
        }
      }
    },
    "/odontologos/{id}": {
      "get": {
        "tags": [
          "Odontólogos"
        ],
        "summary": "Buscar odontólogo por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do odontólogo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "$ref": "#/components/schemas/Odontologo"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Odontólogo não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Odontólogos"
        ],
        "summary": "Atualizar cadastro de odontólogo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OdontologoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cadastro atualizado"
          }
        }
      },
      "delete": {
        "tags": [
          "Odontólogos"
        ],
        "summary": "Excluir cadastro de odontólogo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Odontólogo excluído"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponseEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Operação executada com sucesso"
              }
            }
          }
        }
      },
...
      "Odontologo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 2 },
          "nome": { "type": "string", "example": "Dr. Fernando Dentista" },
          "cro": { "type": "string", "example": "TEST-1234" },
          "salario": { "type": "number", "format": "float", "example": 12000.00 },
          "id_especialidade": { "type": "integer", "example": 1 },
          "id_usuario": { "type": "integer", "example": 10 }
        }
      },
      "OdontologoInput": {
        "type": "object",
        "required": [
          "nome",
          "cro",
          "salario",
          "id_especialidade",
          "id_usuario"
        ],
        "properties": {
          "nome": { "type": "string", "example": "Dr. Fernando Dentista" },
          "cro": { "type": "string", "example": "TEST-1234" },
          "salario": { "type": "number", "minimum": 0, "example": 12000.00 },
          "id_especialidade": { "type": "integer", "example": 1 },
          "id_usuario": { "type": "integer", "example": 10 }
        }
      },
...
    }
  }
}